+++ core/services/chainlink/application.go
@@ -293,7 +293,7 @@ func NewApplication(opts ApplicationOpts) (Application, error) {
 		Config:                   cfg,
 		webhookJobRunner:         webhookJobRunner,
 		KeyStore:                 keyStore,
-		SessionReaper:            services.NewSessionReaper(db, cfg),
+		SessionReaper:            sessions.NewSessionReaper(opts.SqlxDB.DB, cfg),
 		Exiter:                   os.Exit,
 		ExternalInitiatorManager: externalInitiatorManager,
 		shutdownSignal:           shutdownSignal,
