+++ core/sessions/reaper_test.go
@@ -1,15 +1,13 @@
-package services_test
+package sessions_test
 
 import (
+	"database/sql"
 	"testing"
 	"time"
 
-	"github.com/smartcontractkit/chainlink/core/internal/cltest"
 	"github.com/smartcontractkit/chainlink/core/internal/testutils/pgtest"
-	"github.com/smartcontractkit/chainlink/core/services"
-	"github.com/smartcontractkit/chainlink/core/services/postgres"
+	"github.com/smartcontractkit/chainlink/core/sessions"
 	"github.com/smartcontractkit/chainlink/core/store/models"
-	"gorm.io/gorm"
 
 	"github.com/onsi/gomega"
 	"github.com/stretchr/testify/assert"
@@ -29,10 +27,11 @@ func (c sessionReaperConfig) ReaperExpiration() models.Duration {
 func TestSessionReaper_ReapSessions(t *testing.T) {
 	t.Parallel()
 
-	db := pgtest.NewGormDB(t)
+	db := pgtest.NewSqlxDB(t)
 	config := sessionReaperConfig{}
+	orm := sessions.NewORM(db, config.SessionTimeout().Duration())
 
-	r := services.NewSessionReaper(db, config)
+	r := sessions.NewSessionReaper(db.DB, config)
 	defer r.Stop()
 
 	tests := []struct {
@@ -50,24 +49,23 @@ func TestSessionReaper_ReapSessions(t *testing.T) {
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
 			t.Cleanup(func() {
-				clearSessions(t, db)
+				clearSessions(t, db.DB)
 			})
 
-			session := cltest.NewSession(test.name)
-			session.LastUsed = test.lastUsed
-			require.NoError(t, db.Save(&session).Error)
+			_, err := db.Exec("INSERT INTO sessions (last_used, id, created_at) VALUES ($1, $2, now())", test.lastUsed, test.name)
+			require.NoError(t, err)
 
 			r.WakeUp()
 
 			if test.wantReap {
-				gomega.NewGomegaWithT(t).Eventually(func() []models.Session {
-					sessions, err := postgres.Sessions(db, 0, 10)
+				gomega.NewGomegaWithT(t).Eventually(func() []sessions.Session {
+					sessions, err := orm.Sessions(0, 10)
 					assert.NoError(t, err)
 					return sessions
 				}).Should(gomega.HaveLen(0))
 			} else {
-				gomega.NewGomegaWithT(t).Consistently(func() []models.Session {
-					sessions, err := postgres.Sessions(db, 0, 10)
+				gomega.NewGomegaWithT(t).Consistently(func() []sessions.Session {
+					sessions, err := orm.Sessions(0, 10)
 					assert.NoError(t, err)
 					return sessions
 				}).Should(gomega.HaveLen(1))
@@ -77,6 +75,7 @@ func TestSessionReaper_ReapSessions(t *testing.T) {
 }
 
 // clearSessions removes all sessions.
-func clearSessions(t *testing.T, db *gorm.DB) {
-	require.NoError(t, db.Exec("DELETE FROM sessions").Error)
+func clearSessions(t *testing.T, db *sql.DB) {
+	_, err := db.Exec("DELETE FROM sessions")
+	require.NoError(t, err)
 }
